stress-ng-1.5.2: change the count number from 0 to -1 in stress-ng
In the function stress_get_processors of stress-ng:
count = 0 -> number of CPUs in system
count < 0 -> number of CPUs online

Using online CPU number will be more accurate if some of the CPU
core are disabled or offline. For example, if we're doing the
scaling performance test, we would like to set the actual online
CPU number as the running instance number in stress-ng.