pts/stress-ng-1.7.0: update to bug fix release of stress-ng 0.15.06
Versions 0.15.04 and 0.15.05 contained some regressions, so update
to the bugfix release 0.15.06.

Update the sock stressor to use the rawsock mode (if it's available)
as this is a minor improvement of the underlying network stack benchmark
that removes the need for a userspace/kernel boundary memory copy per
send/recv and is more in keeping with modern performant socket
utilization patterns.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>