pts/cloverleaf-1.1.0: set optimial GOMP_SPINCOUNT for large systems
For large systems of over 80 CPUs there is a large amount of memory
contention in gomp_team_barrier_wait_end that impacts on performance
and unnecessarily increases run time. This can be resolved by setting
GOMP_SPINCOUNT to 7000 to reduce the contention. Also ensure that
the OMP_WAIT_POLICY is set to active to ensure the optimial OMP wait
policy is used for larger systems.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>